+++ services/s3/src/test/java/software/amazon/awssdk/services/s3/S3UtilitiesTest.java
@@ -29,7 +29,6 @@ import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
 import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.s3.model.GetUrlRequest;
-import software.amazon.awssdk.services.s3.parsing.S3Uri;
 
 public class S3UtilitiesTest {
